requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
pydantic
httpx